Fixed tests
